added docs for new WikiReader, renamed module accordingly, and updated imports
