+++ textacy/corpora/__init__.py
@@ -1,3 +1,3 @@
-from .wikipedia import WikiReader
+from .wiki_reader import WikiReader
 from .reddit_reader import RedditReader
 from .bernie_and_hillary import fetch_bernie_and_hillary
